Examiner corrections
Implemented the corrections from the examiners.

Examiner 1
1. p6,9 Indicate source of Figure 2.1, 2.4
(response)For visually qualitative reasons, all images in the
dissertation, that are non-photographic, are drawn by me (except the
UKZN logo). These two figures are a combination of many similar figures
that have tried to visually depict the physical process of fluorescence
and microscopy. These exact figures cannot be found anywhere else, but
is one among many variants. Citing a single source would be difficult
given the multiplicity and the subtle nuances of images elsewhere.
2. p10 Check the appropriateness of the statement “Chapter 8 takes a
step in this vein” in Section 2.4
3. p11 Replace the statement “It is due the interaction…” by “It is due
to the interaction…”
4. p12 The operator in equation (2.1) should be defined.
5. p15 Replace the statement “This technique widely used…” by “This
technique is widely used…”
6. p16 Check the appropriateness of the statement “Differential chain
codes, and its two most common descriptors, boundary chain code (BCC)
and differential chain code (DCC)…”
7. p25 Replace the statement “we are interested with cutting…” by “we
are interested in cutting…”
8. p69 Replace the statement “…we can see that the all methods…” by
““…we can see that all methods…”
9. p142 Replace the statement “Although the weighting scheme presented
by…” by “Although the weighting schemes presented by…”
10. p166 Replace the statement “we achieved an average…” by “We achieved
an average…”
11. p167 Replaced the statement “The proposed scheme show an…” by “The
proposed scheme shows an…”
12. p189 The year of publication is missing in reference 195 on the
reference list.

Examiner 2
1. pxxxiii In the list of abbreviations, the full form of EMGMM should
be included.  Done
2.  In some of the references the year of publication is mentioned in
brackets i.e. (2005) and in some references, it is without brackets i.e.
2005. It should be uniform for all references.
(response) I agree with the examiner with regards to the question of
consistency of the format. This can be seen by comparing ref-6 and
ref-7. However, a citation style of “(AIP) - American Institute of
Physics” has been used, and this is automatically generated by LaTex.
Furthermore, given the time constraints, amending the format could
jeopardise the situation even further and could possibly result in a
non-conventional citation standard.
3. p137 The word ‘aproximation’ should be corrected to ‘approximation’.
Done
4. p70,169 The value 1.1391% should be 1.1397%.